+++ SME_Implementations/SingleCycleRISCV/Mem_mux_Buses.cs
@@ -5,7 +5,7 @@ namespace SingleCycleRISCV {
     [TopLevelOutputBus]
     public interface Mem_Mux_Output : IBus {
         [InitialValue]
-        int Data {get; set;}
+        long Data {get; set;}
     }
 
 }
