+++ SME_Implementations/SingleCycleRISCV/Register.cs
@@ -21,7 +21,7 @@ namespace SingleCycleRISCV {
         
 
         // private readonly int[] m_register = new int[32];
-        private readonly int[] m_register = {0, 0, 0, 0, 0, 0, 0, 0,
+        private readonly long[] m_register = {0, 0, 0, 0, 0, 0, 0, 0,
                                              0, 0, 0, 0, 0, 0, 0, 2,
                                              8, 7, 0, 0, 0, 0, 0, 0,
                                              0, 0, 0, 0, 0, 0, 0, 0,};
